Rename 'Element' to 'Unit' for Tensor and ShapedArray.
TL;DR: 'Unit' will replace 'Element' as a denotation of the type of atom data (scalars) in a Tensor or ShapedArray. 'Element' will instead denote the type of the sub-dimensional (rank-1) aggregate in a Tensor or ShapedArray.

In a flattened collection, the associated type 'Element' often denotes the type of values that the collection is made of. However, Tensor introduced a shaped world to Swift collections, where one can index a tensor using an integer to get the subdimensioanl tensor. Even though Tensor is not itself a collection, using the word 'Element' for the atom data type of a tensor is no longer consistent with Swift's conventional indexing APIs which almost always yield the atom data type. Moreover, ShapedArrays, which will be designed as a Collection of lower-ranked arrays, will have name conflicts due to Collection's associated type 'Element'. This means that the current 'Element' must be renamed to something else.

Before renaming 'Element', We need to redefine what 'Element' really means in a shaped world. The answer is clear: 'Element' is not the atom data type in a Tensor or a ShapedArray, but the shaped mathematical object's sub-dimensional aggregate. As for the atom data type, 'Unit' is a perfectly suitable name, formally defined as "an individual thing regarded as single and complete but which can also form an individual component of a larger or more complex whole".

Changes are the following:
- Rename '.contiguousView' to '.units'.
- Rename ShapedArray's '.contiguousView' to 'units'.
- Rename 'TensorElementProtocol' to 'TensorAccelerable', denoting data types that can be accelerated and parallelized in a Tensor.
- Rename Array's '.totalElementCount' to '.unitCount'.
- Rename Tensor's '.elements' to '.array', as 'array' has much better indication of type conversion (and memory transfer in our API).

PiperOrigin-RevId: 182006809