[ClangImporter] Anonymous enums used as types can't be imported as Int
Normally we treat anonymous enums as defining constants, and since we
don't know how those constants are going to be used we import them as
Int when we know we can do so consistently across all platforms. But
this isn't the right thing to do if the anonymous enum is being used
as an ad-hoc type for a variable or field.

(This may /still/ not be the right thing to do---maybe we want to
import it with a made-up name like we do anonymous fields instead.
But at least this way any such variables or fields become usable,
when they previously crashed IRGen.)

https://bugs.swift.org/browse/SR-2511