WIP: DO NOT MERGE: force optional value
In https://github.com/apple/swift/pull/13842 I reverted this change
due to a source compatibility failure, but unfortunately I didn't note
which test failed and the logs are gone. So I'm opening this PR in
order to run the source compat suite again so that I can extract a
small test case showing what fails with this change.