Sema: validate class layout when validating an extension
This reverts a change done in 9b7f3f73d2c8368c5db0ba0f978ad960c39e9970, in a rare case, shown in rdar://problem/36704036, When doing emitVirtualMethodValue in IRGen we can miss the correct offset because we have a bad ClassDecl, The ClassDecl is junk / null because of the before mentioned commit.

The radar contains a sample ClassDecl and IR that shows the problem, unfortunately, I am unable to create a small test-case the reproduces this issue: the layout is wrong only in a single method call in a large project. Reducing the project to something of a reasonable size / removing most of the code “hides” the problem.

Basically, if we look at the ClassDecl, we see that the type for some vars is null type (v.s. the correct type), the interface type is missing and we are missing a few accessor_decl  and parameter_list