SILGen: Make the optionality check in checkForABIDifferences reflexive to make up for the loss of IUO.
This function was also trying to play the role of covariance police, which doesn't really seem appropriate to me, and relying on the difference between IUO and Optional to decide whether it was OK to consider an [ImplicitlyUnwrapped]Optional<T> to be compatible with T. Sometimes we allow the "wrong" direction due to IUO-ness, or due to conservative nullness in ObjC bridging lowering. It's really the type checker's job to make sure the right thing happens at the semantic level, so we should be safe here in making the check here symmetric instead of directional. Potentially fixes rdar://problem/36843476.