[StringGuts] Bias the unmanaged pointer by StringStorage's header.
By storing a biased pointer (e.g. -32 on 64-bit), we can make the
"native" and "unmanaged" cases layout-identical except for the isValue
bit. This means that string literals are basically strings that are
immutable and immortal.

Removes some branching when trying to convert to an unmanaged
view. This reduces the size of unmanagedASCIIView (frequently inlined
into user code) by about 1/3rd.

Idea came through conversation with Jordan Rose. Thanks @jrose-apple