Turn StrideTo into a slice and conditionally conform to Collection
StrideTo used to be a sequence, and a very minimalistic one, and so the
majority of Sequence methods and properties it obtained from the
protocol extensions. One other thing that it got from those default
implemetnations is SubSequence == AnySequence, which is not particularly
good when we consider conditionally confiorm it to Collection.

This commit contains the necessary changes to make StrideTo become a "slice
type", i.e. StrideTo.SubSequence is now StrideTo. Conditional
conformance for the case of integer elements is also implemented, it
should be modified, however, as now indices of StrideTo and
StrideTo.SubSequence are not interchangeable, which is expected from the
standard library collection types.