build: depend on the architecture variant on non-MachO
For the non-MachO targets, which do not have the concept of fat
binaries, depend explicitly on the architecture variant.  This is
particularly useful for Windows which permits cross-compilation and
cross-targeting of all the variants even though it does not support fat
binaries.  This allows us to build a single variant of the windows
standard library at a time in the case that we would like to build a
subset of the targets.