build: strip the lipo'ed target placeholder
When building non-MachO targets, there is no support for fat binaries.
As a result, the lipo target does nothing.  Because we do not support
variant builds for ELF targets, this has not resulted in any issues.
However, the PE/COFF (Windows) target suppots multiple variants.  In
such a case, the "lipo" target will copy a number of builds, leaving the
binary to be an arbitrary variant.  Instead, prefer to always use the
architectural variant path.  In place, create an empty target which
serves as a means of collecting dependencies.