performTypeChecking: Defer verifyAllLoadedModules in WMO mode.
Suggested by Jordan Rose.

Defer per-file verification in whole-module mode. Verifying imports between
files could cause the importer to cache declarations without adding them to the
ASTContext. This happens when the importer registers a declaration without a
valid TypeChecker instance, as is the case during verification. A subsequent
file may require that declaration to be fully imported (e.g. to synthesized a
function body), but since it has already been cached, it will never be added to
the ASTContext. The solution is to skip verification and avoid caching
it. Instead, loaded modules are now verified in WMO mode once during
performWholeModuleTypeChecking.

Fixes <rdar:36801676> [Edge][swift 4.1] SIL verification failed: external
declarations of SILFunctions with shared visibility is not allowed.