[test] Added a missing import to test/stdlib/Reflection_objc.swift.
`import CoreGraphics` is required for this test on non-macOS Apple platforms. (On macOS, `import Foundation` is sufficient.)
This previously was not an issue, as this test used to `import UIKit` on non-macOS Apple platforms.
Since CoreGraphics is available everywhere this test runs, and this test does test CoreGraphics types, I just added an import of CoreGraphics.