stdlib: use `llvm::function_ref` over `std::function`
This is needed to repair the Windows build.  The MSVC C++ runtime
implements `std::function` with use of `typeid` which cannot be used
with `-fno-rtti`.  This results in a build failure.  Rather than
enabling the use of RTTI for `MetadataLookup.cpp`, use
`llvm::function_ref` to avoid enabling RTTI emission.