Added a more thorough fix to balance retain/release SIL instructions after host code rewrite.
Design summary: when removing a builtin tensor op instruction TO, for each
tensor argument TA of it, if TA is NOT generated by other tensor ops, emit a
strong_release to balance the already emitted strong_retain's for TA. Recall the
calling convention of TO is that TO consumes TA at +1.

For example, before removing instruction:
  %r1 = builtin "__tfop_Add__tt:t__"(%arg1, %arg2),
we first emit strong_release'es on %arg1 and %arg2.

If there is a subsequent instruction:
  %r2 = builtin "__tfop_Add__tt:t__"(%r1, %arg3),
we only emit strong_release on %arg3, and not on %r1, since the strong_retain on
%r1 will be removed when all references of the first Add instruction are dropped
along with that instruction itself.

Added a few unit tests.

PiperOrigin-RevId: 184065965