SIL: use `unsigned` for the member for `bool`
Use `unsigned` instead of `bool` for the member to make all members of
the structure the same.  This fixes the behaviour of the structure on
Windows.  The difference in the type caused the bitfields to be pushed
to the natural alignment of int resulting in the structure being padded
to 12-bytes instead of being packed into 4.  Using `#pragma pack` also
did not make a difference here as the structure remained padded to
12-bytes.