Make DiagnoseUnreachable and NoReturnFolding function passes.
We want as few module passes as possible.

Function passes allow the PassManager to do its job.
e.g. it can filter certain functions that should not be applied to the
current pipeline. This will result in less work in the pass itself and
fewer pass manager related bugs.

Function passes are easier to understand and debug in the context of the
pipeline. Things like PrettyStackTrace are handled automatically.
Bisecting functionality is builtin.

Function passes are more compatible in general with inter-procedural
analysis.

Function passes are more efficient.

A single module pass in the middle of the pipeline destroys the benefit
of the rest of the pipeline uses function passes.