Fix AllocBoxToStack to notify the PassManager of cloned functions.
If the pass does not register a new function, that new function will never be
passed through the current function pipeline. We were getting lucky that enough
of the mandatory passes were module passes that it didn't seem to matter in
practice (until I convert them to function passes).

(after pass reorganization, a one line fix that doesn't rely on
outrageous changes to the pass' function signatures.)