Add SILFunction::WasDeserialized flag.
We need to be able to detect function definitions that have been
deserialized. There's no need to rerun diagnostics on those functions,
and in some cases it's actually incorrect to do so. With exclusivity,
we could even miscompile in theory (debug assert) because the fully
optimized SIL does adhere to rules requires by the
diagnostic. Hopefully that specific issue can be fixed soon, but the
point is that we need control over the order that passes are run
because we play these games all the time. Not to mention the wasted
compilation time.

isAvailableExternally probably won't work reliably. What if we
deserialize an always-emit-into-client definition? It would not have
external linkage.

I would strongly prefer not to rely on this flag for correctness. In
principle, serialized SIL should be compatible with all SIL stages
prior to the serialization point. However, it is tremendously useful
for bootstrapping and reducing compile time, and is necessary to fix
bugs in the short term.