Add ModuleFile::isSIB
.sib files are not modules in any sense. They obey a different set of rules from
serialized Swift modules. However, we currently depend use the ModuleFile
abstraction to serialize/deserialize .sib. Until we have a better abstraction,
the ModuleFile will need to know whether it's a serialized Swift module or
actually just a .sib file.

Alternatively, we could serialize the module's stage, and we probably should do
so for .sib.