Introduce the ability to represent string an array attributes in #tfop syntax, e.g. this now works:
  func convolved2D(withFilter filter: Tensor, strides: [Int], padding: Padding) -> Tensor {
    return Tensor(#tfop("Conv2D", "tt:t", handle, filter.handle,
                        strides: strides, padding: padding.cName))
  }

This requires a bunch of different things to work out.  First we have to
pattern match the SIL code that builds the array so we can recognize the
initialization and check for constants, this is what decodeArrayElements does.

Then we have to canonicalize the representation so all the array SIL code is
out of the way.  We transform this initial internal representation:

  %45 = builtin "__tfop_Conv2D,tt:t,strides,padding"(%37 : $TensorHandle<Float>, %4 : $TensorHandle<Float>, %19 : $Array<Int>, %44 : $String) : $TensorHandle<Float>

into this, which inlines all of the elements (this is all internal to the
compiler, no user of #tfop ever sees this):

  %46 = builtin "__tfop_Conv2D,tt:t,strides$array,$elt,$elt,$elt,$elt,padding"(%37 : $TensorHandle<Float>, %4 : $TensorHandle<Float>, %45 : $@thin Int.Type, %21 : $Builtin.Int64, %26 : $Builtin.Int64, %31 : $Builtin.Int64, %6 : $Builtin.Int64, %38 : $Builtin.RawPointer) : $TensorHandle<Float>

This is handled by the changes in SILTensorOpInfo::canonicalizeOperands.

The last big piece is the change to GraphLowering, which needs to turn this
representation into the right TensorFlow C API calls to build the lists of
various sorts.

There are still a few things missing: we don't support the array form of
$tensor, don't support $shape, and don't support arrays of dtypes yet.
I'll add these whenever they become a priority and have a testcase.

PiperOrigin-RevId: 184358121