Add CanType::getOptionalObjectType for staging a rename.
In order to replace uses of getAnyOptionalObjectType() with
getOptionalObjectType(), we need to update LLDB. But in order
to do that, we need to have the version of this function for
CanType available.

So this adds this, as a temporary measure, to facilitate the LLDB
update that will unblock the compiler rename.