Really fix b/72908345 by fixing how we delete SILArguments that get moved to the accelerator.  BB arguments have a fixed numbering between the branches that feed values to the block and the block arguments.  When we go to remove the values that feed the branches, we would break this numbering for subsequent arguments of the same block.
Resolve this by handling all bb arguments of a given block at once, so we can
keep track of the association between the argument number we're deleting and
the arguments to delete.

This adds the failing test to the crashers suite in unreduced form.  It isn't a
great reduced testcase, but having some coverage there is useful too.  The
runtime test still doesn't work (complaining about lack of send support).  I'll
fix the problems affecting .mean() next.

PiperOrigin-RevId: 184557101