WIP: Teach SwiftSyntax to use SourceKit to serialize syntax trees.
When using SwiftSyntax as a standalone tool, we invoke Swiftc
internally to get the serialized syntax tree. This is not ideal for
several reasons: (1) we have to hard-code the relative path of swiftc
to invoke it; (2) we have to rely on standard input/output to pass the
tree across the process boundaries; and (3) we have to maintain two
different ways to get syntax tree (swiftc and sourcekitd).

This patch attempts to teach SwiftSyntax to use SourceKit to get the
tree just like other clients. We first add a SourceKit client library
written in Swift; and next teach SwiftSyntax to use this SourceKit
client-side library. For platforms other than MacOS, we still use Swiftc
to get syntax trees.

This patch also re-enables several flaky tests.