Fix transpose and reverse ops, add tests.
The `reversed()` method is fixed: it now uses "ReverseV2" instead of the "Reverse", which has a less intuitive API.
Argumentless `transposed()` is fixed for Tensor/TensorXD.
To produce the default dimension permutations for transpose, the Tensor version uses `rankTensor` to construct a range, then reverses it.
The TensorXD version efficiently constructs a reversed range since rank is statically known.

PiperOrigin-RevId: 184790020